refactor: Change test package structure
