+++ src/main/kotlin/model/graph_model/graph_model_actions/VertViewUpdate.kt
@@ -2,7 +2,7 @@ package model.graph_model.graph_model_actions
 
 import androidx.compose.ui.graphics.Color
 
-class VertViewUpdate(
+data class VertViewUpdate(
     val color: Color? = null,
     val alpha: Float? = null
 )
