+++ src/test/kotlin/model/graph_modelTests/GraphTest.kt
@@ -1,7 +1,7 @@
-package model.graph_model
+package model.graph_modelTests
 
+import model.graph_model.Graph
 import org.junit.jupiter.api.Assertions.*
-import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import kotlin.test.assertFailsWith
