+++ src/test/kotlin/model/graph_modelTests/UndirectedGraphTest.kt
@@ -1,5 +1,6 @@
-package model.graph_model
+package model.graph_modelTests
 
+import model.graph_model.UndirectedGraph
 import org.junit.jupiter.api.Test
 
 import org.junit.jupiter.api.Assertions.*
