+++ src/test/kotlin/model/graph_modelTests/algorithmTest/BridgeFindingTest.kt
@@ -1,4 +1,4 @@
-package modelTest
+package model.graph_modelTests.algorithmTest
 
 import androidx.compose.runtime.mutableStateMapOf
 import androidx.compose.ui.graphics.Color
